Move arg parsing to util
